Extract cache tests from unit test collection
